Move `HeaderView` to a separated file
